+++ index.html
@@ -7,7 +7,7 @@
   </head>
 
   <body>
-    <div id="app"></div>
+    <div class="app" id="app"></div>
   </body>
   <script type="module" src="./utils.js"></script>
   <script type="module" src="./main.js"></script>
