+++ styles.css
@@ -1,7 +1,9 @@
 body {
     margin: 0;
   }
-
+  .app {
+    background: #202020;
+  }
   .container {
     font-family: Helvetica;
     color: #ffffff;
